requirements.txt
# Cloud SDKs
tencentcloud-sdk-python==3.0.1070
alibabacloud_bssopenapi20171214==1.2.6
alibabacloud_tea_openapi==0.3.9
alibabacloud_tea_util==0.3.12

# Data processing and ML
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
prophet==1.1.5

# Django
Django==4.2.7

# Database
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
requests==2.31.0
